pyproject.toml
[project]
name = "gn-mcp"
version = "0.1.0"
description = "Glassnode API MCP Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]",
    "httpx",
    "python-dotenv",
]